Make data types of currency values consistent to `decimal(18, 8)`
This is due to an issue related to using `float` datatype being prone to errors. See #213 .

+++ database/migrations/2022_03_10_012743_create_taxes_table.php
@@ -17,7 +17,7 @@ class CreateTaxesTable extends Migration
             $table->id();
             $table->foreignId('accounting_system_id')->constrained();
             $table->string('name');
-            $table->float('percentage');
+            $table->decimal('percentage', 18, 8);
             $table->timestamps();
         });
     }
